+++ platforms/node/src/index.d.ts
@@ -173,6 +173,7 @@ export declare class MermanError extends Error {
 }
 
 export interface MermanResourceErrorDetails {
+  readonly cause: string;
   readonly limit_id: string;
   readonly phase: string;
   readonly actual: number;